+++ tests/conftest.py
@@ -1,6 +1,6 @@
 from typing import Any, Dict, Optional
 
-from pytest import Session
+import pytest
 
 
 def _patch_plotly_show() -> None:
@@ -23,15 +23,21 @@ def _patch_plotly_show() -> None:
     plotly.io.show = wrapped
 
 
-def pytest_sessionstart(session: Session) -> None:
-    """Called after the :py:class:`pytest.Session` object has been created and
+def pytest_sessionstart(session: pytest.Session) -> None:
+    """Called after the :class:`pytest.Session` object has been created and
     before performing collection and entering the run test loop.
 
-    Args:
-        session
-            The pytest :py:class:`~pytest.Session` object.
+    Parameters
+    ----------
+    session
+        The pytest :class:`~pytest.Session` object.
 
-    References:
-    - https://docs.pytest.org/en/6.2.x/reference.html#initialization-hooks
+    References
+    ----------
+    https://docs.pytest.org/en/stable/reference.html#initialization-hooks
+
+    ..
+        Ignore the following `flake8-docstrings` errors:
+        # noqa: D401
     """
     _patch_plotly_show()