+++ CHANGES.md
@@ -12,6 +12,22 @@ Unreleased changes
 - 🔧 Fixed and improved some type annotations, including the introduction of `ridgeplot._types`
   module for type aliases such as `Numeric` and `NestedNumericSequence`.
   ([#80](https://github.com/tpvasconcelos/ridgeplot/pull/80))
+- 🔧 Add the `blacken-docs` pre-commit hook and add the `pep8-naming`, `flake8-pytest-style`,
+  `flake8-simplify`, `flake8-implicit-str-concat`, `flake8-bugbear`, `flake8-rst-docstrings`,
+  `flake8-rst-docstrings`, etc... plugins to the `flake8` pre-commit hook.
+  ([#81](https://github.com/tpvasconcelos/ridgeplot/pull/81))
+- 📝 Change to numpydoc style docstrings.
+  ([#81](https://github.com/tpvasconcelos/ridgeplot/pull/81))
+- 📝 Add a robots.txt to the docs site.
+  ([#81](https://github.com/tpvasconcelos/ridgeplot/pull/81))
+- 📝 Auto-generate a site map for the docs site using `sphinx_sitemap`.
+  ([#81](https://github.com/tpvasconcelos/ridgeplot/pull/81))
+- 📝 Change the sphinx theme to `furo`.
+  ([#81](https://github.com/tpvasconcelos/ridgeplot/pull/81))
+- 📝 Improve the internal documentation and some of these internals to the API reference.
+  ([#81](https://github.com/tpvasconcelos/ridgeplot/pull/81))
+- 💅 Cleanup and improve some type annotations.
+  ([#81](https://github.com/tpvasconcelos/ridgeplot/pull/81))
 
 0.1.20
 ------